Create separated file to export a project
